third_party/WebKit/LayoutTests/external/wpt/media-source/mediasource-addsourcebuffer-mode-expected.txt

This is a testharness.js-based test.
PASS addSourceBuffer() sets SourceBuffer.mode to 'segments' when the generate timestamps flag is false
FAIL addSourceBuffer() sets SourceBuffer.mode to 'sequence' when the generate timestamps flag is true assert_equals: expected "sequence" but got "segments"
Harness: the test ran to completion.
